Ensure connections are closed if missing sufficient responses after timeout.
I can't see anywhere that a timed out query reduces the number of replies that must be read before a connection at the query limit will close and re-open. So keep track of the number of queries in flight, decrement that when a reply is received or on a timeout (we don't want to decrement the number of messages received, surely?) and close when the in flight count goes to 0.